+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0
   - Upgrade baseimage: light-baseimage:0.2.0
+  - Makefile with build no cache
 
 ## 0.1.9
   - Upgrade baseimage: light-baseimage:0.1.5